btc_oracle/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btc-oracle"
version = "0.1.0"
description = "BTCUSDT Adaptive Oracle - прогнозирование без торговли"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "BTC Oracle Team"}
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.3.0",
    "ta-lib>=0.4.0; sys_platform != 'win32'",
    "ta>=0.11.0",
    "structlog>=23.2.0",
    "python-dateutil>=2.8.2",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "lmdb>=1.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[project.scripts]
titan = "btc_oracle.cli:cli"
btc-oracle-live = "btc_oracle.app:main"
btc-oracle-api = "btc_oracle.api.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

